[BB-1245] Move changes after merging base branch
